+++ CHANGELOG.md
@@ -16,11 +16,14 @@ All user visible changes to this project will be documented in this file. This p
 - UI:
     - Home page:
         - Wallet and monetization tabs. ([#2])
+    - Wallet tab:
+        - Top up and transactions pages. ([#3])
 - Deployment:
     - [Helm] chart. ([#1])
 
 [#1]: /../../pull/1
 [#2]: /../../pull/2
+[#3]: /../../pull/3